fix method substs
